+++ src/index.js
@@ -1,1 +1,3 @@
+import './scss/index.scss';
+
 export * from './webgis4u';