+++ src/webgis4u/ol/control/common.js
@@ -5,7 +5,7 @@
 /**
  * The default CSS class that OL uses for controls
  */
-export const CSS_CONTROL_DEFAULT = 'ol-control';
+export const OL_CSS_CONTROL_DEFAULT = 'ol-control';
 /**
  * The default CSS class that OL uses for contols
  * that can't be selected
@@ -13,6 +13,14 @@ export const CSS_CONTROL_DEFAULT = 'ol-control';
 export const CSS_CONTROL_UNSELECTABLE = 'ol-unselectable';
 
 /**
+ * General prefix for CSS class selectors
+ */
+export const CSS_PREFIX = 'webgis4u';
+/**
  * Prefix used for controls
  */
-export const CSS_CONTROL_PREFIX = 'webgis4u-control';
+export const CSS_CONTROL_PREFIX = `${CSS_PREFIX}-control`;
+
+export const OL_CONTROL_ZOOM_CSS_CLASS_MOUSE_POSITION = 'ol-mouse-position';
+export const OL_CONTROL_ZOOM_CSS_CLASS_ZOOM = 'ol-zoom';
+export const OL_CONTROL_ZOOM_CSS_CLASS_ZOOMSLIDER = 'ol-zoomslider';