+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 * Added projections for the austrian area - [`austria`](./src/webgis4u/ol/proj/austria.js)
 * Added [`basemap`](./src/webgis4u/ol/source/basemap.js) as a source - for more information, see [basemap.at](https://www.basemap.at/)
 * Added following controls:
+  * [`HtmlTemplate`](./src/webgis4u/ol/control/HtmlTemplate.js)
   * [`PanBar`](./src/webgis4u/ol/control/PanBar.js)
   * [`LayerCheckbox`](./src/webgis4u/ol/control/LayerCheckbox.js)
   * [`LayerRadioButton`](./src/webgis4u/ol/control/LayerRadioButton.js)
@@ -14,13 +15,16 @@ All notable changes to this project will be documented in this file.
   * [`OverviewMap`](./src/webgis4u/ol/control/OverviewMap.js)
   * [`ScaleLine`](./src/webgis4u/ol/control/ScaleLine.js)
 * Added utils:
+  * [`dom`](./src/webgis4u/util/dom.js):
+    * [`toggleElement`](./src/webgis4u/util/dom/toggleElement.js)
   * [`function`](./src/webgis4u/util/function.js):
     * [`debounce`](./src/webgis4u/util/function/debounce.js):
   * [`number`](./src/webgis4u/util/number.js):
     * [`convertUnitValue`](./src/webgis4u/util/number/convertUnitValue.js):
   * [`ol`](./src/webgis4u/util/ol.js):
     * [`getMapChildElementByClassName`](./src/webgis4u/util/ol/getMapChildElementByClassName.js):
-    * [`pan`](./src/webgis4u/util/o#l/pan.js):
+    * [`getMapType`](./src/webgis4u/ol/util/getMapType.js)
+    * [`pan`](./src/webgis4u/util/ol/pan.js)
   * [`string`](./src/webgis4u/util/string.js):
     * [`formatNumber`](./src/webgis4u/util/string/formatNumber.js):
     * [`formatUnitValue`](./src/webgis4u/util/string/formatUnitValue.js):